% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba.R
\name{ba}
\alias{ba}
\title{Basal Area - m2/ha}
\usage{
ba(DBH, plotsize)
}
\arguments{
\item{DBH}{Field containing the tree DBH in cm}

\item{plotsize}{Area of the plot in hectares}
}
\value{
The basal area in
}
\description{
The basal area is the area of a given section of land that is occupied by the cross-section of tree trunks and stems at the base.
}
\examples{
\dontrun{
ba(DBH = dbh_2,plotsize = 0.02)
}
}
