% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_taper.R
\name{bi2000_squared}
\alias{bi2000_squared}
\title{Taper bi2000 squared}
\usage{
bi2000_squared(DBH, h, H, coefficients)
}
\arguments{
\item{DBH}{Tree DBH in cm}

\item{h}{Stem diameter height in m}

\item{H}{Tree height in m}

\item{coefficients}{equation coefficients}
}
\value{
squared stem diameter at h
}
\description{
Internal function to be integrated for the integration of the Bi 2000 volume function
}
\examples{
\dontrun{
bi2000(30,5,40,coef())
}
}
