% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_taper.R
\name{bi2000_volume}
\alias{bi2000_volume}
\title{Volume bi2000}
\usage{
bi2000_volume(DBH, stump_height, h, H, coefficients)
}
\arguments{
\item{DBH}{Tree diameter at breast height in cm}

\item{stump_height}{Tree stump height in m}

\item{h}{Height for predicted diameter}

\item{H}{Tree height in m}

\item{coefficients}{Equation coefficients}
}
\value{
Stem volume
}
\description{
D Bi, H; 2000; Trigonometric variable-form taper equations for Australian Eucalypts; Forest Science 46(3) pp397-409.
}
\examples{

bi2000_volume(30,0.3, 40,40,coefficients = bi_tapercoef_aus('CMAC'))


}
