% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{clearfellAge_line}
\alias{clearfellAge_line}
\title{Line graph showing clearfell ages}
\usage{
clearfellAge_line(data, Year, value, colourVariable, WAF_form)
}
\arguments{
\item{data}{Dataframe containing woodflow outputs. This will typically have been produced using e.g. Remsoft Woodstock}

\item{Year}{Column in df containing values for the x-axis}

\item{value}{Column containing values to be plotted on the y-axis. This will typically be volume}

\item{colourVariable}{Variable to use for assigning fill colour to the bar graph}

\item{WAF_form}{Boolean indicating whether plot for WAF reporting or not. Default = TRUE}
}
\description{
Function for plotting line graphs of clearfell ages from Woodflow model outputs
}
