% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLE.R
\name{conf.int}
\alias{conf.int}
\title{Confidence interval}
\usage{
conf.int(mu)
}
\arguments{
\item{mu}{The sampled variable. This might commonly be set of values calculated from inventory plots}
}
\value{
The confidence limit in the same units as mu
}
\description{
Calculate the confidence limit for a confidence interval around a mean at the 95\% confidence level
}
\examples{
vols<-runif(20, 150, 250)
conf.int(mu = vols)

}
