% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{is_outlier}
\alias{is_outlier}
\title{Identify outliers}
\usage{
is_outlier(x, type = "both", outlier_range = 1.5)
}
\arguments{
\item{x}{The variable of interest}

\item{type}{Choose whether to identify 'high', 'low', or 'both' types of outliers. Default is 'both'.}

\item{outlier_range}{The number of times outside the IQR that an observation is deemed to be an outlier. Default = 1.5.}
}
\description{
Function that identifies whether observations are outliers.
}
\examples{
x <- rnorm(n = 1000, mean = 100) # Create a normally distributed dataset

is_outlier(x) # Find outliers

boxplot(x) # Plot x to visualise the outliers
}
