% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{is_outlier_high}
\alias{is_outlier_high}
\title{Identify high outliers}
\usage{
is_outlier_high(x, outlier_range = 1.5)
}
\arguments{
\item{x}{The variable of interest}

\item{outlier_range}{The number of times outside the IQR that an observation is deemed to be an outlier. Default = 1.5}
}
\description{
Function that identifies whether observations are high outliers.
}
\examples{
x <- rnorm(n = 1000, mean = 100)
is_outlier_high(x)
}
