% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{MagGroome_cols}
\alias{MagGroome_cols}
\title{Extract Margules Groome colour hex codes}
\usage{
MagGroome_cols(...)
}
\arguments{
\item{...}{Character name of MagGroome_colors}
}
\value{
hex code for Margules Groome Colour
}
\description{
Function to extract Margules Groome colours as hex codes
}
\examples{
MagGroome_cols('mgcblue')

}
