% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{MagGroome_pal}
\alias{MagGroome_pal}
\title{Return function to interpolate a Mag Groome color palette}
\usage{
MagGroome_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in MagGroome_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a Mag Groome color palette
}
\examples{
MagGroome_pal('main') (10)

}
