% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDD.R
\name{mdd}
\alias{mdd}
\title{MDD}
\usage{
mdd(DBH, H, plotsize)
}
\arguments{
\item{DBH}{Field containing the tree DBHs in cm}

\item{H}{Field containing the tree heights in m}

\item{plotsize}{Plot area in ha}
}
\value{
Mean dominant DBH (cm)
}
\description{
Calculate the plot mean dominant DBH
}
\examples{
dbh<-runif(20, 20 ,30)
h<-runif(20, 25, 35)
plot.size <- 0.06
mdd(DBH = dbh, H = h, plotsize = plot.size)
}
