% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTH.R
\name{mth}
\alias{mth}
\title{Mean Top Height}
\usage{
mth(DBH, H, plotsize)
}
\arguments{
\item{DBH}{Field containing the tree DBH values in cm measured in the plot}

\item{H}{Field containing the tree height measurements in m.}

\item{plotsize}{Area of the plot in hectares}
}
\value{
The mean top height of the plot.
}
\description{
The average height of the primary leaders of the largest 100 trees per hectare where largest is measured in terms of the dbh of the primary leader of that tree.
}
\examples{
\dontrun{
mth(DBH = dbh_m, H = h_m, plotsize = 0.05)
}
}
