% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDH.R
\name{pdh}
\alias{pdh}
\title{Predominant Mean Height}
\usage{
pdh(DBH, H, plotsize)
}
\arguments{
\item{DBH}{Tree DBH in cm}

\item{H}{Total tree height in m.}

\item{plotsize}{Plot area in ha}
}
\value{
The predominant top height of the plot.
}
\description{
The average height of the primary leaders of the largest 100 trees per hectare where largest is measured in terms of the height of the primary leader of that tree.
}
\examples{
\dontrun{
mth(DBH, H, plotsize)
}
}
