% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLE.R
\name{ple}
\alias{ple}
\title{Probably limit of error (PLE) around a mean}
\usage{
ple(mu)
}
\arguments{
\item{mu}{The sampled variable. This might commonly be set of values calculated from inventory plots}
}
\value{
The PLE in percent
}
\description{
Calculate the PLE around a mean at the 95\% confidence level. This statistic is commonly used in forest inventory analysis in NZ and Australia
}
\examples{
vols<-runif(20, 150, 250)
ple(mu = vols)

}
