% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{prepare_WAF_allrep}
\alias{prepare_WAF_allrep}
\title{Prepare WAF allrep data}
\usage{
prepare_WAF_allrep(df, startYear, cutOverYear)
}
\arguments{
\item{df}{The WAF allrep file.}

\item{startYear}{Simulation start year.}

\item{cutOverYear}{simulation cutover year.}
}
\description{
Function that transforms data from the allrep output of Remsoft Woodstock
}
