% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMD.R
\name{QMD}
\alias{QMD}
\title{Quadratic mean DBH}
\usage{
QMD(DBH)
}
\arguments{
\item{DBH}{tree DBH values}
}
\value{
Quadratic mean DBH
}
\description{
Calculate the quadratic mean DBH
}
\examples{
\dontrun{
QMD(DBH = dap)
}
}
