% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(predicted, actual)
}
\arguments{
\item{predicted}{Predicted value}

\item{actual}{Actual values}
}
\value{
Root mean square error
}
\description{
Calculate the RMSE
}
