% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{scale_colour_MagGroome}
\alias{scale_colour_MagGroome}
\title{Color scale constructor for MagGroome colors}
\usage{
scale_colour_MagGroome(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in MagGroome_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for MagGroome colors
}
\examples{
scale_colour_MagGroome()


}
