% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{scale_fill_MagGroome}
\alias{scale_fill_MagGroome}
\title{Fill scale constructor for MagGroome colours}
\usage{
scale_fill_MagGroome(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in MagGroome_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for MagGroome colours
}
\examples{
scale_fill_MagGroome()

}
