% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{shift_legend}
\alias{shift_legend}
\title{Shift legend}
\usage{
shift_legend(p)
}
\arguments{
\item{p}{The ggplot2 multi-panel object.}
}
\description{
Shift legend into empty facets in a multipanel ggplot.
}
