% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spha.R
\name{spha}
\alias{spha}
\title{SPHA}
\usage{
spha(DBH, plotsize)
}
\arguments{
\item{DBH}{Field containing the tree DBH in cm}

\item{plotsize}{Area of the plot in hectares}
}
\value{
Stems per hectare with a DBH greater than 0
}
\description{
Return stems per hectare
}
\examples{
\dontrun{
spha(DBH = dap,plotsize = 0.02)
}
}
