% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLE.R
\name{std.err}
\alias{std.err}
\title{Standard error}
\usage{
std.err(mu)
}
\arguments{
\item{mu}{The sampled variable. This might commonly be set of values calculated from inventory plots}
}
\value{
The standard error in the same units as mu
}
\description{
Calculate the standard error of a sample
}
\examples{
vols<-runif(20, 150, 250)
std.err(mu = vols)
}
