% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpha.R
\name{tpha}
\alias{tpha}
\title{TPHA}
\usage{
tpha(TreeNo, plotsize)
}
\arguments{
\item{TreeNo}{Tree name}

\item{plotsize}{Plot size in hectares}
}
\value{

}
\description{
Function to return the number of trees per hectare
}
