% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAndYield_Functions.R
\name{tpr_max_burkhart_safh}
\alias{tpr_max_burkhart_safh}
\title{Stem taper}
\usage{
tpr_max_burkhart_safh(DBH, H, h, coefficients)
}
\arguments{
\item{DBH}{Tree DBH in cm}

\item{H}{Tree height in m}

\item{h}{Stem height}

\item{coefficients}{Equation coefficients}
}
\value{
stem diameter at height h
}
\description{
Max and Burkhart taper function with regression coefficients from the
South African forestry handbook
}
