% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAndYield_Functions.R
\name{vol_tpr_coef_sa}
\alias{vol_tpr_coef_sa}
\title{SA Vol TPR coefficients}
\usage{
vol_tpr_coef_sa(species)
}
\arguments{
\item{species}{Species code}
}
\value{
Returns equation coefficients for South Africa.
}
\description{
Coefficients for the volume and taper equation from the South African Forestry handbook
}
\examples{
vol_tpr_coef_sa('EGRA')
}
