% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAndYield_Functions.R
\name{vol_tree_safh}
\alias{vol_tree_safh}
\title{Estimate tree volume}
\usage{
vol_tree_safh(
  DBH,
  H,
  species = "E. grandis",
  b0 = -9.746,
  b1 = 1.715,
  b2 = 1.107,
  f = -2
)
}
\arguments{
\item{DBH}{Tree DBH in cm}

\item{H}{Tree height in m}

\item{species}{Tree species}

\item{b0}{Equation coefficient}

\item{b1}{Equation coefficient}

\item{b2}{Equation coefficient}

\item{f}{Form factor modifier from handbook}
}
\value{
Returns tree volume in cubic metres
}
\description{
Estimate tree volume from DBH and height using functions from the South
African Forestry Handbook. These species are available. 'A. mearnsii', 'A. melanoxylon', 'C. equisetifolia', 'E. camaldulensis', 'E. cladocalyx', 'E. diversicolor', 'E. fastigata', 'E. globulus', 'E. grandis', 'E. maculata', 'E. nitens', 'E. grandis x camaldulensis', 'P. canariensis', 'P. caribea', 'P. elliotti', 'P. patula', 'P. pinaster', 'P. radiata', 'P. taeda', 'T. grandis'
}
