% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{volume_graph_bar}
\alias{volume_graph_bar}
\title{Bar graph showing woodflow volumes}
\usage{
volume_graph_bar(data, value, Year, Species_select, fillvariable)
}
\arguments{
\item{data}{Dataframe containing woodflow outputs. This will typically have been produced using e.g. Remsoft Woodstock}

\item{value}{Column containing values to be plotted on the y-axis. This will typically be volume}

\item{Year}{Column in df containing values for the x-axis}

\item{Species_select}{Which species is being plotted}

\item{fillvariable}{Variable to use for assigning fill colour to the bar graph}
}
\description{
Function for plotting bar graphs of volumes from Woodflow model outputs
}
