% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafR_functions.R
\name{volume_graph_line}
\alias{volume_graph_line}
\title{Line graph showing woodflow volumes}
\usage{
volume_graph_line(
  data,
  value,
  Year,
  colourvariable,
  WAF_form = TRUE,
  Species_select = "Radiata_pine",
  ylim = 7e+05
)
}
\arguments{
\item{data}{Dataframe containing woodflow outputs. This will typically have been produced using e.g. Remsoft Woodstock}

\item{value}{Column containing values to be plotted on the y-axis. This will typically be volume}

\item{Year}{Column in df containing values for the x-axis}

\item{colourvariable}{Variable to use for assigning fill colour to the bar graph}

\item{WAF_form}{True/False indicating whether plot for WAF reporting or not. Default = TRUE}

\item{Species_select}{Which species is being plotted if this is a WAF report output, Default = 'Radiata_pine'}

\item{ylim}{Maximum value for the y axis}
}
\description{
Function for plotting line graphs of volumes from Woodflow model outputs
}
