% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba.R
\name{ba}
\alias{ba}
\title{Basal Area - m2/ha}
\usage{
ba(DBH, plotsize)
}
\arguments{
\item{DBH}{Field containing the tree DBH in cm}

\item{plotsize}{Area of the plot in hectares}
}
\value{
The plot basal area in m2/ha
}
\description{
The basal area is the area of a given section of land that is occupied by the cross-section of tree stems at breast height.
}
\examples{

dbh = runif(20, 25, 35) # Make an example set of dbhs
ba(DBH = dbh,plotsize = 0.02)

}
