% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAndYield_Functions.R
\name{tpr_max_burkhart_safh}
\alias{tpr_max_burkhart_safh}
\title{Stem taper}
\usage{
tpr_max_burkhart_safh(DBH, H, h, species, a1, a2, b5, b6, b7, b8)
}
\arguments{
\item{DBH}{Tree DBH in cm}

\item{H}{Tree height in m}

\item{h}{Stem height}

\item{species}{Tree species}

\item{a1}{Equation coefficient}

\item{a2}{Equation coefficient}

\item{b5}{Equation coefficient}

\item{b6}{Equation coefficient}

\item{b7}{Equation coefficient}

\item{b8}{Equation coefficient}
}
\value{
stem diameter at height h
}
\description{
Max and Burkhart taper function with regression coefficients from the
South African forestry handbook
}
